// src/types/error-codes.ts
// enum - перечисляемый тип, который позволяет создавать именнованный набор констант.
export enum ErrorCodes {
	UNHANDLED = "UNHANDLED",
	EXISTS = "EXISTS",
	NOT_FOUND = "NOT_FOUND",
	UNAUTHORIZED = "UNAUTHORIZED",
	VALIDATION = 'VALIDATION',
}